apps/crawl4ai-domain-monitor/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
pydantic-settings==2.1.0
python-multipart==0.0.6

# Crawl4AI library - main dependency
crawl4ai==0.3.74

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# HTTP client
httpx==0.25.2

# Background tasks
celery==5.3.4
redis==5.0.1

# Utilities
python-dateutil==2.8.2
feedparser==6.0.10
beautifulsoup4==4.12.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
psutil==5.9.6